[feat][jw-w2] VM 테스트 코드 수정
- VM 테스트 코드 수정
- PokemonList package 이동
